+++ service/src/test/java/com/softserve/edu/service/calibrator/data/test/impl/CalibrationTestServiceImplTest.java
@@ -150,7 +150,7 @@ public class CalibrationTestServiceImplTest {
         Verification.ConsumptionStatus status = Verification.ConsumptionStatus.IN_THE_AREA;
         ArgumentCaptor<String> nameArg = ArgumentCaptor.forClass(String.class);
         when(testRepository.findOne(testId)).thenReturn(calibrationTest);
-        CalibrationTest calibrationTest = calibrationTestService.editTest(testId, name, 1, 1, 1d, 1d, status,
+        CalibrationTest calibrationTest = calibrationTestService.editTest(testId, name, 1, 1L, 1d, 1d, status,
                 Verification.CalibrationTestResult.SUCCESS);
         verify(calibrationTest).setName(nameArg.capture());
         assertEquals(name, nameArg.getValue());
@@ -205,7 +205,7 @@ public class CalibrationTestServiceImplTest {
     public void testCreateNewCalibrationTest() throws Exception {
         String name = "name";
         Integer temperature = 20;
-        Integer settingNumber = 10;
+        Long settingNumber = 10L;
         Double latitude = 24d;
         Double longitude = 48d;
         CalibrationTest expected = new CalibrationTest();
